+++ frontend/components/common/board/types.ts
@@ -1,0 +1,9 @@
+import { ApplicantPassState } from "../../../src/apis/kanban";
+
+export interface BoardData {
+  id: string;
+  title: string;
+  subElements: string[];
+  time?: Date;
+  passState?: ApplicantPassState;
+}
